requirements.txt
python-dotenv
vaderSentiment
fpdf
wordcloud